seek(): allow seeking beyond EOF
This change is perhaps a bit controversial, but given that rejecting
seeks that exceed EOF is not standardized across languages supported in
Kaitai Struct, I think it might be better to just ignore this case for
now. It seems more future-proof: if we start throwing some arbitrary
error now, chances are we'll have to change it later for consistency
with other languages once it is standardized, which might break backward
compatibility because some Rust user might already depend on the
specific format of the temporary error.

I believe most languages allow seeks beyond EOF at the moment. Some
don't (e.g. Java) because their standard library doesn't allow that, so
we'll perhaps reach the conclusion that we'll indeed throw a
standardized error everywhere (since that would be the lowest common
denominator), but such decision has not been made yet.