Remove unused `KError::UnexpectedContents` variant
It is not used from the rest of the `src/lib.rs` file and it is not
referenced from
https://github.com/kaitai-io/kaitai_struct_compiler/pull/250 either.
Besides, it is apparently a reproduction of `UnexpectedDataError`, which
was used in KS only before 0.9. Since 0.9, `ValidationNotEqualError` is
used instead (see
https://github.com/kaitai-io/kaitai_struct/issues/280#issuecomment-544238045).
In runtime libraries which existed before 0.9, `UnexpectedDataError` was
marked as deprecated and will be eventually removed, so it doesn't make
sense to add it to new runtime libraries.