Add `KError::ValidationFailed`, remove `KError::ValidationNotEqual`
At the time of writing,
https://github.com/kaitai-io/kaitai_struct_compiler/pull/250 used
`KError::ValidationNotEqual` for all kinds of validations (see
https://github.com/kaitai-io/kaitai_struct_compiler/blob/23328785803ba93a8b3bfc33e1e0b998a34df01f/shared/src/main/scala/io/kaitai/struct/languages/RustCompiler.scala#L1199),
which doesn't make any sense.

This commit makes a step towards a solution consistent with the
established system of validation errors in KS. The enum variant
`KError::ValidationFailed` represents the failure of any validation (it
is an equivalent of the `ValidationFailedError` base class in other
languages). It contains a `ValidationFailedError` struct, which is
intended to hold any details associated with the specific failed
validation. It only contains the `kind` field at the moment, but the
plan is to extend it in the future with fields like `actual`, `src_path`
etc.

The `ValidationKind` enum has been marked with the [`non_exhaustive`
attribute](https://doc.rust-lang.org/reference/attributes/type_system.html#the-non_exhaustive-attribute),
which allows us to add more validation kinds in the future without
breaking compatibility with user application code.

Even though I anticipate the addition of fields to the
`ValidationFailedError` struct, it cannot be marked `non_exhaustive`
because we must be able to instantiate it from the KSC-generated code.
So I added at least a doc comment warning the users about the
possibility of extending the struct contents.