Allow bytes processing routines to be fallible
This is needed to fix the (very valid) warning that we ignored a
`Result` value in the `process_zlib` method (this makes sense, because
`process: zlib` can definitely fail if the input is not a valid zlib
stream, for example if it doesn't have a valid zlib header). It makes
sense to allow custom (user-defined) processing routines to fail as
well.