perf: move downcast out of trait
it is only used internally and should not be shared as much.
While removing this from the trait is technically breaking its not changing the behaviour and isnt used by ksc so it can be changed relatively easily.

This allows it to change the arguments reducing the amount of clones.